Подключил пакет day, формат времени
